tests: simplify code by renaming packed charms and dirs to just substrate
