+++ pathops/tests/integration/juju/charms/machine/charmcraft.yaml
@@ -1,4 +1,4 @@
-name: test-machine
+name: machine
 type: charm
 summary: A small machine charm for use in integration tests.
 description: A small machine charm for use in integration tests.
